Add files via upload
The Dataset contains the data for the 100 top-rated movies from the past decade along with various pieces of information about the movie, its actors, and the voters who have rated these movies online. In this assignment, The objective is to find some interesting insights into these movies and their voters, using Python.